+++ functions/src/validators/auth.validator.ts
@@ -1,0 +1,5 @@
+import { header } from 'express-validator';
+
+export const authValidator = [
+  header('authorization').isJWT().notEmpty().withMessage('Missing OAuthToken'),
+];