+++ src/app/components/loading-overlay/loading-overlay.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-loading-overlay',
+  standalone: true,
+  imports: [],
+  templateUrl: './loading-overlay.component.html',
+  styleUrl: './loading-overlay.component.scss',
+})
+export class LoadingOverlayComponent {}