+++ src/app/pages/about-me/credentials/credentials.component.ts
@@ -1,10 +1,10 @@
 import { Component, Input, OnInit, Renderer2 } from '@angular/core';
-import { MicrosoftLearnUserProfile } from 'src/app/interfaces/credentials/microsoft/microsoft';
+import { MicrosoftLearnUserProfile } from 'src/app/interfaces/credentials/microsoft/microsoft.interface';
 import { Tabs } from 'src/app/interfaces/tabs.model';
 import { SafeHtml } from '@angular/platform-browser';
 import credInfo from 'src/assets/credentials/msft_credentials.json';
 import { credentials } from 'src/assets/credentials/credly_credentials';
-import { Credly } from 'src/app/interfaces/credentials/credly/credly';
+import { Credly } from 'src/app/interfaces/credentials/credly/credly.interface';
 import {
   MatCard,
   MatCardContent,
@@ -65,7 +65,7 @@ export class CredentialsComponent extends Tabs implements OnInit {
     this.renderer2.appendChild(targeto, scriptEl);
 
     this.profile = new MicrosoftLearnUserProfile(credInfo);
-    console.log(this.profile);
+    console.debug(this.profile);
   }
   @Input() tabTitle: string;
 }