+++ src/app/services/link-preview/link-preview.service.spec.ts
@@ -12,11 +12,12 @@ import {
 import { connectAuthEmulator, getAuth, provideAuth } from '@angular/fire/auth';
 import { environment } from 'src/environments/environment';
 import sinon from 'sinon';
-import { AppCheckToken, getToken } from 'firebase/app-check';
+import { AuthService } from 'src/app/services/auth-service/auth.service';
 // const quibble = require('quibble');
 
 describe('LinkPreviewService', () => {
   let service: LinkPreviewService;
+  let authService: AuthService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -42,6 +43,7 @@ describe('LinkPreviewService', () => {
       ],
     });
     service = TestBed.inject(LinkPreviewService);
+    authService = TestBed.inject(AuthService);
   });
 
   afterEach(() => {
@@ -51,7 +53,7 @@ describe('LinkPreviewService', () => {
 
   it('should be created with appCheck service', () => {
     expect(service).toBeTruthy();
-    expect(service.appCheck).toBeDefined();
+    // expect(service.appCheck).toBeDefined();
   });
 
   // it('should get an app check token', async () => {